// packages/rill/src/types.ts
export interface CreateOperation extends BaseOperation {
  op: 'CREATE';
  type: string;
  props: SerializedProps;
}

/**
 * Update node operation
 */
export interface UpdateOperation extends BaseOperation {
  op: 'UPDATE';
  props: SerializedProps;
  removedProps?: string[];
}

/**
 * Append child node operation
 */
export interface AppendOperation extends BaseOperation {
  op: 'APPEND';
  parentId: number;
  childId: number;
}

/**
 * Insert child node operation
 */
export interface InsertOperation extends BaseOperation {
  op: 'INSERT';
  parentId: number;
  childId: number;
  index: number;
}

/**
 * Delete node operation
 */
export interface DeleteOperation extends BaseOperation {
  op: 'DELETE';
}

/**
 * Remove child node operation
 */
export interface RemoveOperation extends BaseOperation {
  op: 'REMOVE';
  parentId: number;
  childId: number;
}

/**
 * Reorder children operation
 */
export interface ReorderOperation extends BaseOperation {
  op: 'REORDER';
  parentId: number;
  childIds: number[];
}

/**
 * Update text operation
 */
export interface TextOperation extends BaseOperation {
  op: 'TEXT';
  text: string;
}

export type Operation =
  | CreateOperation
  | UpdateOperation
  | AppendOperation
  | InsertOperation
  | DeleteOperation
  | RemoveOperation
  | ReorderOperation
  | TextOperation;
// ...
export interface OperationBatch {
  version: number;
  batchId: number;
  operations: Operation[];
}
// ...
export interface NodeInstance {
  id: number;
  type: string;
  props: Record<string, unknown>;
  children: number[];
}